mlolongo genaral hospital version3  learning csv file creation, reading and appending(adding data)

+++ Templates/csv file writer template.py
@@ -1,7 +1,11 @@
 import csv
+import os
 headers=["Name","Age","Ward","Diagnosis","Gender","DOA"]
-file=open("New_patients.csv","w",newline="")
-writer=csv.writer(file)
-writer.writerow(headers)
-file.close()
-print("File created with headers successfully")
+if os.path.exists("New_patients.csv"):
+    print("File for patients already exists, creation skipped")
+else:
+    file=open("New_patients.csv","w",newline="")
+    writer=csv.writer(file)
+    writer.writerow(headers)
+    file.close()
+    print("File created with headers successfully!")
